Changed a lot:
separated code into packages. The GA package will hold everything related to the ga. The neural_network package will hold the structure of the neural net. The purpose of this is to allow other classes to operate on the network structure.
The structure will be accessible through the Network.java class.

I still need to do:
-building Connection objects for the network structure. I am still thinking about how to store them.
-Training (GD).

I will try to do these later today.

Signed-off-by: cauth0n <derek.reimanis@gmail.com>