Fixed an issue with mutation and changed some other things. Somehow it's still booting the best individual, that or it's fitness is getting evaluated differently during different runs.
All in all, it runs better. We're consistently getting fitness in the 90s, and it looks like it peaks out around 94.5%, a better fitness than backprop with it's current error threshold.

I'm thinking we should definitely save a copy of the best individul for when we exit the GA process, in case we've somehow decreased the max fit individual.